[#1018] Separate domain & status cell from table
