+++ src/components/DomainsTable.vue
@@ -1,15 +1,13 @@
 <script setup lang="ts">
-  import DomainVerificationInfo from '@/components/DomainVerificationInfo.vue';
-  import DomainsTableRowActionDropdown from '@/components/dashboard/DomainsTableRowActionDropdown.vue';
+import DomainsTableDomainCell from '@/components/dashboard/DomainsTableDomainCell.vue';
+  import DomainsTableActionsCell from '@/components/dashboard/DomainsTableActionsCell.vue';
   import ToggleWithIcon from '@/components/common/ToggleWithIcon.vue';
   import type { CustomDomain } from '@/schemas/models/domain';
-  import { useDomainStatus } from '@/composables/useDomainStatus';
-  import { computed } from 'vue';
+
   import OIcon from '@/components/icons/OIcon.vue';
-  import { formatDistanceToNow } from 'date-fns';
   import { useBranding } from '@/composables/useBranding';
 
-  const props = defineProps<{
+  defineProps<{
     domains: CustomDomain[];
     isLoading: boolean;
   }>();
@@ -18,13 +16,6 @@
     (e: 'delete', domainId: string): void;
     (e: 'toggle-homepage', domain: CustomDomain): void;
   }>();
-
-  const domainStatuses = computed(() =>
-    props.domains.map((domain) => ({
-      domain,
-      status: useDomainStatus(domain),
-    }))
-  );
 
   const handleDelete = (domainId: string) => {
     emit('delete', domainId);
@@ -100,52 +91,12 @@
         </thead>
 
         <tbody class="divide-y divide-gray-200 bg-white dark:divide-gray-700 dark:bg-gray-900">
-          <tr v-for="{ domain, status } in domainStatuses"
+          <tr v-for="domain in domains"
               :key="domain.identifier"
               class="transition-colors duration-150 hover:bg-gray-50 dark:hover:bg-gray-800">
             <!-- Domain & Status -->
             <td class="px-6 py-4">
-              <div class="flex flex-col">
-                <div class="flex items-center gap-2">
-                  <router-link v-if="status.isActive.value"
-                               :to="{ name: 'DomainBrand', params: { domain: domain.display_domain } }"
-                               class="font-brand text-lg text-brandcomp-600 hover:text-brandcomp-700 dark:text-brandcomp-400 dark:hover:text-brandcomp-300">
-                    {{ domain.display_domain }}
-                  </router-link>
-                  <router-link v-else
-                               :to="{
-                      name: 'DomainVerify',
-                      params: { domain: domain.display_domain },
-                    }"
-                               class="font-brand text-lg text-brandcomp-600 hover:text-brandcomp-700 dark:text-brandcomp-400 dark:hover:text-brandcomp-300">
-                    {{ domain.display_domain }}
-                  </router-link>
-                  <a :href="`https://${domain.display_domain}`"
-                     target="_blank"
-                     rel="noopener noreferrer"
-                     class="text-gray-400 hover:text-gray-600 dark:text-gray-500 dark:hover:text-gray-300"
-                     title="Open domain in new tab">
-                    <svg xmlns="http://www.w3.org/2000/svg"
-                         class="h-4 w-4"
-                         viewBox="0 0 24 24"
-                         fill="none"
-                         stroke="currentColor">
-                      <path stroke-linecap="round"
-                            stroke-linejoin="round"
-                            stroke-width="2"
-                            d="M10 6H6a2 2 0 00-2 2v10a2 2 0 002 2h10a2 2 0 002-2v-4M14 4h6m0 0v6m0-6L10 14" />
-                    </svg>
-                  </a>
-                </div>
-
-                <div class="mt-1 flex items-center gap-2">
-                  <DomainVerificationInfo mode="icon"
-                                          :domain="domain" />
-                  <span class="text-xs text-gray-500 dark:text-gray-400">
-                    Added {{ formatDistanceToNow(domain.created, { addSuffix: true }) }}
-                  </span>
-                </div>
-              </div>
+              <DomainsTableDomainCell :domain="domain" />
             </td>
 
             <!-- Homepage Access -->
@@ -160,8 +111,8 @@
 
             <!-- Actions -->
             <td class="px-6 py-4 text-right">
-              <DomainsTableRowActionDropdown :domain="domain"
-                                             @delete="handleDelete" />
+              <DomainsTableActionsCell :domain="domain"
+                                       @delete="handleDelete" />
             </td>
           </tr>
         </tbody>
