+++ src/components/dashboard/DomainsTableActionsCell.vue
